# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UNZA DigiHealth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Users
...
  /consultations:
    get:
      operationId: listConsultations
      tags: [consultations]
      summary: List consultations
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ConsultationStatus"
        - name: studentId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of consultations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Consultation"

    post:
      operationId: createConsultation
      tags: [consultations]
      summary: Create a new consultation request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConsultationRequest"
      responses:
        "201":
          description: Created consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}:
    get:
      operationId: getConsultation
      tags: [consultations]
      summary: Get consultation by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Consultation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"

    put:
      operationId: updateConsultation
      tags: [consultations]
      summary: Update consultation (doctor response/diagnosis)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationRequest"
      responses:
        "200":
          description: Updated consultation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"
  /consultations/{id}/status:
    patch:
      operationId: updateConsultationStatus
      tags: [consultations]
      summary: Update consultation status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConsultationStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Consultation"

  # Queue
...
  /hiv-support/sessions:
    get:
      operationId: listHivSupportSessions
      tags: [hiv-support]
      summary: List HIV support sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HivSupportSession"

    post:
      operationId: createHivSupportSession
      tags: [hiv-support]
      summary: Request a HIV support session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHivSupportSessionRequest"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HivSupportSession"
  /hiv-support/sessions/{id}:
    put:
      operationId: updateHivSupportSession
      tags: [hiv-support]
      summary: Update HIV support session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHivSupportSessionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HivSupportSession"
  /hiv-support/resources:
    get:
      operationId: listHivResources
      tags: [hiv-support]
      summary: List HIV/AIDS educational resources
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Resources list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HivResource"

  # Dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    UserRole:
      type: string
      enum: [student, doctor, pharmacist, lab_technician, mental_health_counselor, hiv_professional, admin]
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
    ConsultationStatus:
      type: string
      enum: [submitted, under_review, assigned, responded, closed]
...
    SessionStatus:
      type: string
      enum: [requested, active, completed, cancelled]

    # Auth
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        studentNumber:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/UserRole"
      required: [name, email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]

    # User
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        studentNumber:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, status, createdAt]
...
    Consultation:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        doctorId:
          type: integer
        symptoms:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        attachments:
          type: array
          items:
            type: string
        status:
          $ref: "#/components/schemas/ConsultationStatus"
        diagnosis:
          type: string
        notes:
          type: string
        student:
          $ref: "#/components/schemas/User"
        doctor:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, studentId, symptoms, severity, status, createdAt, updatedAt]
    CreateConsultationRequest:
      type: object
      properties:
        symptoms:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        attachments:
          type: array
          items:
            type: string
      required: [symptoms, severity]
    UpdateConsultationRequest:
      type: object
      properties:
        diagnosis:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/ConsultationStatus"

    # Queue
...
    HivSupportSession:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        professionalId:
          type: integer
        topic:
          type: string
        notes:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        appointmentDate:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
        professional:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required: [id, studentId, topic, status, createdAt]
    CreateHivSupportSessionRequest:
      type: object
      properties:
        topic:
          type: string
        appointmentDate:
          type: string
          format: date-time
        notes:
          type: string
      required: [topic]
    UpdateHivSupportSessionRequest:
      type: object
      properties:
        notes:
          type: string
        status:
          $ref: "#/components/schemas/SessionStatus"
        appointmentDate:
          type: string
          format: date-time
    HivResource:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, content, category, createdAt]

    # Dashboard
...
    UpdateConsultationStatusRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ConsultationStatus"
        doctorId:
          type: integer
      required: [status]
...