# lib/api-zod/src/generated/types/hivResource.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UNZA DigiHealth API
 * OpenAPI spec version: 0.1.0
 */

export interface HivResource {
  id: number;
  title: string;
  content: string;
  category: string;
  createdAt: Date;
}