# lib/api-zod/src/generated/types/userStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UNZA DigiHealth API
 * OpenAPI spec version: 0.1.0
 */

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;